# src/main/resources/META-INF/openapi.json
{
    "openapi": "3.0.2",
    "info": {
        "title": "Adress Book API",
        "version": "1.0.0",
        "license": {
            "name": "Apache 2.0",
            "url": "https://www.apache.org/licenses/LICENSE-2.0"
        }
    },
    "paths": {
        "/addresses": {
            "get": {
                "responses": {
                    "200": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Address"
                                    }
                                }
                            }
                        },
                        "description": "Returns a list of addresses"
                    }
                },
                "summary": "Get list of addresses"
            },
            "post": {
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Address"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "201": {
                        "description": "Created response"
                    }
                },
                "summary": "Creating a new address"
            }
        },
        "/addresses/{addressId}": {
            "get": {
                "responses": {
                    "200": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Address"
                                }
                            }
                        },
                        "description": "Returned address by id"
                    }
                },
                "summary": "Get address by id"
            },
            "put": {
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Address"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "Address updated response"
                    }
                },
                "summary": "Update address"
            },
            "delete": {
                "responses": {
                    "204": {
                        "description": "Address deleted"
                    }
                },
                "summary": "Delete address by id"
            },
            "parameters": [
                {
                    "name": "addressId",
                    "schema": {
                        "type": "string"
                    },
                    "in": "path",
                    "required": true
                }
            ]
        }
    },
    "components": {
        "schemas": {
            "Address": {
                "title": "Root Type for Address",
                "description": "",
                "type": "object",
                "properties": {
                    "id": {
                        "format": "int32",
                        "type": "integer"
                    },
                    "country": {
                        "type": "string"
                    }
                },
                "example": {
                    "id": 1,
                    "country": "ES"
                }
            }
        }
    }
}